requirements.txt
flask
flask-login
flask-wtf
flask-bcrypt
python-dotenv
pandas
matplotlib
plotly
scipy
pytest
